Refactor Docker image
