+++ Makefile
@@ -4,21 +4,60 @@
 #
 
 PROJECT_NAME             := metricfunc
-VERSION                  ?= $(shell cat ./VERSION)
+DOCKER_VERSION           ?= $(shell cat ./VERSION)
 
 ## Docker related
 DOCKER_REGISTRY          ?=
 DOCKER_REPOSITORY        ?=
-DOCKER_TAG               ?= ${VERSION}
+DOCKER_TAG               ?= ${DOCKER_VERSION}
 DOCKER_IMAGENAME         := ${DOCKER_REGISTRY}${DOCKER_REPOSITORY}${PROJECT_NAME}:${DOCKER_TAG}
 DOCKER_BUILDKIT          ?= 1
 DOCKER_BUILD_ARGS        ?=
 
+## Docker labels. Only set ref and commit date if committed
+DOCKER_LABEL_VCS_URL     ?= $(shell git remote get-url $(shell git remote))
+DOCKER_LABEL_VCS_REF     ?= $(shell git diff-index --quiet HEAD -- && git rev-parse HEAD || echo "unknown")
+DOCKER_LABEL_COMMIT_DATE ?= $(shell git diff-index --quiet HEAD -- && git show -s --format=%cd --date=iso-strict HEAD || echo "unknown" )
 DOCKER_LABEL_BUILD_DATE  ?= $(shell date -u "+%Y-%m-%dT%H:%M:%SZ")
 
 DOCKER_TARGETS           ?= metricfunc
 
-# https://docs.docker.com/engine/reference/commandline/build/#specifying-target-build-stage---target
+GO_BIN_PATH = bin
+GO_SRC_PATH = ./
+C_BUILD_PATH = build
+ROOT_PATH = $(shell pwd)
+
+NF = $(GO_NF)
+GO_NF = metricfunc
+
+NF_GO_FILES = $(shell find $(GO_SRC_PATH)/$(%) -name "*.go" ! -name "*_test.go")
+
+VERSION = $(shell git describe --tags)
+BUILD_TIME = $(shell date -u +"%Y-%m-%dT%H:%M:%SZ")
+COMMIT_HASH = $(shell git submodule status | grep $(GO_SRC_PATH)/$(@F) | awk '{print $$(1)}' | cut -c1-8)
+COMMIT_TIME = $(shell cd $(GO_SRC_PATH) && git log --pretty="%ai" -1 | awk '{time=$$(1)"T"$$(2)"Z"; print time}')
+
+.PHONY: $(NF) clean docker-build docker-push
+
+.DEFAULT_GOAL: nfs
+
+nfs: $(NF)
+
+all: $(NF)
+
+$(GO_NF): % : $(GO_BIN_PATH)/%
+
+$(GO_BIN_PATH)/%: %.go $(NF_GO_FILES)
+# $(@F): The file-within-directory part of the file name of the target.
+	@echo "Start building $(@F)...."
+	cd $(GO_SRC_PATH)/ && \
+	CGO_ENABLED=0 go build -o $(ROOT_PATH)/$@ $(@F).go
+
+vpath %.go $(addprefix $(GO_SRC_PATH)/, $(GO_NF))
+
+clean:
+	rm -rf $(addprefix $(GO_BIN_PATH)/, $(GO_NF))
+	rm -rf $(addprefix $(GO_SRC_PATH)/, $(addsuffix /$(C_BUILD_PATH), $(C_NF)))
 
 docker-build:
 	@go mod vendor
@@ -41,7 +80,25 @@ docker-push:
 		docker push ${DOCKER_REGISTRY}${DOCKER_REPOSITORY}$$target:${DOCKER_TAG}; \
 	done
 
+.coverage:
+	rm -rf $(CURDIR)/.coverage
+	mkdir -p $(CURDIR)/.coverage
 
-.PHONY: docker-build docker-push
+test: .coverage
+	docker run --rm -v $(CURDIR):/metricfunc -w /metricfunc golang:latest \
+		go test \
+			-race \
+			-failfast \
+			-coverprofile=.coverage/coverage-unit.txt \
+			-covermode=atomic \
+			-v \
+			./ ./...
 
+fmt:
+	@go fmt ./...
 
+golint:
+	@docker run --rm -v $(CURDIR):/app -w /app golangci/golangci-lint:latest golangci-lint run -v --config /app/.golangci.yml
+
+check-reuse:
+	@docker run --rm -v $(CURDIR):/metricfunc -w /metricfunc omecproject/reuse-verify:latest reuse lint
